Refactor the code to clean it up a bit, and also fix the bugs to get things working
Study the diff carefully to see what I fixed, and the improvements I made.
You had two bugs:
* The first was your ternary for comp_team. One half was evaluating fine, but the other half was evaluating to a boolean.
* The second bug was the user_move function was returning a new set, instead of just a tuple to add to a set.
I went ahead and added docstrings to all your functions as well. Try to get in the habit of doing this when you first create a function, it will help you vocalize what that function is actually supposed to accomplish so that you don't lose track of it later.